machine-issue-solver/requirements.txt
# LangGraph
langgraph

# LangChain Core
langchain-core
langchain-community

# Utilities
python-dotenv
requests
pydantic
streamlit